requirements.txt
-r telegram/requirements.txt

Jinja2==3.1.6
MarkupSafe==3.0.2
python-decouple==3.8
setuptools==80.7.1
wheel==0.45.1
